// ffvv/projects/portal/src/app/features/kpis/models/ui/retention-ui.interface.ts
export interface IRetentionUI {
  campaign?: string;
  section?: string;
  i1d1?: number;
  i2d2?: number;
  i3d3?: number;
  i4d4?: number;
  i5d5?: number;
  i6d6?: number;
}

export interface IRetentionsUI {
  section?: string;
  i1d1?: string;
  i2d2?: string;
  i3d3?: string;
  i4d4?: string;
  i5d5?: string;
  i6d6?: string;
}